DuplicateAttribute: Remove the duplicate force resource attribute
Add descriptions for the resource attributes.
Bump the cookbook version.
Add the name_attribute to the resource name attribute.
Add a test case to verify the use of the name attribute.
Include a stripped down version of kitchen_aide.
The kitchen_aide cookbook is used for test kitchen testing of Solaris.
Turn off rubocop complexity alert in the provider.